Apply Double type on seekTo() and getCurrentTime()
- milliseconds 계산을 위해 int에서 double로 type 변경

+++ android/src/main/java/com/inprogress/reactnativeyoutube/YouTubePlayerController.java
@@ -158,7 +158,7 @@ public class YouTubePlayerController implements
     }
 
     public void seekTo(double second) {
-        if (isLoaded()) mYouTubePlayer.seekToMillis(second * 1000);
+        if (isLoaded()) mYouTubePlayer.seekToMillis(second * (double)1000);
     }
 
     public double getCurrentTime() {
